integrate CNN into test_ppo_rl_module

+++ rllib/algorithms/ppo/ppo_rl_module_base.py
@@ -30,9 +30,11 @@ class PPORLModuleBase(RLModule, abc.ABC):
         catalog.encoder_config.input_dim = self.config.observation_space.shape[0]
         catalog.pi_head_config.input_dim = catalog.encoder_config.output_dim
         if isinstance(self.config.action_space, gym.spaces.Discrete):
-            catalog.pi_head_config.output_dim = self.config.action_space.n
+            catalog.pi_head_config.output_dim = int(self.config.action_space.n)
         else:
-            catalog.pi_head_config.output_dim = self.config.action_space.shape[0] * 2
+            catalog.pi_head_config.output_dim = int(
+                self.config.action_space.shape[0] * 2
+            )
         catalog.vf_head_config.output_dim = 1
 
         # Build models from catalog
